Changed alias and token models slightly
In Alias, removed alias id and made a composite primary key on alias and
agent id. Agent->alias = one->many, so this is the simplest structure
and I don't think we care if more than one agent have the same alias.
In Token, added autoincrement to token_id (so we can just insert tokens
without checking for next id).